Reorganise reader/writer tests
