+++ jena-arq/src/test/java/org/apache/jena/riot/TS_RiotGeneral.java
@@ -35,11 +35,6 @@ import org.junit.runners.Suite.SuiteClasses ;
     , TestRDFParser.class
     , TestRDFWriter.class
     , TestParseURISchemeBases.class
-
-    , TestTurtleTrigWriter.class
-    , TestTurtleWriterPretty.class
-
-    , TestJsonLDReader.class
 })
 
 public class TS_RiotGeneral
